Add files via upload
Analyzing Socioeconomic Indicators in Chicago data-set with SQL and Python
The city of Chicago released a dataset of socioeconomic data to the Chicago City Portal. This dataset contains a selection of six socioeconomic indicators of public health significance and a “hardship index,” for each Chicago community area, for the years 2008 – 2012.

Scores on the hardship index can range from 1 to 100, with a higher index number representing a greater level of hardship.

Downloadable report or details of the dataset is available from Census data City of Chicago on their the city of Chicago's website

In this study I have stored a real world data set from the City of Chicago dataset about their scoring on socioeconomic in a database, gain insights into data using SQL queries and then also visualized a portion of the data in the database to see what story it tells.